requirements.txt
# Core dependencies
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.0.0
lightgbm>=3.3.0
ccxt>=2.0.0
aiohttp>=3.8.0
pyyaml>=6.0.0

# Solana-specific
solana>=0.27.0
base58>=2.1.0
construct>=2.10.0
typing-extensions>=4.0.0

# Utilities
python-dateutil>=2.8.2
pytz>=2022.1
requests>=2.28.0
websockets>=10.0